+++ src/index.ts
@@ -1,5 +1,6 @@
 export * from './ContainerImpl';
 export * from './dependencies';
+export * from './errors';
 export * from './inject';
 export * from './singleton';
 export * from './types';